+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -14,6 +14,7 @@
 * Empty strings are now detected as invalid language codes in the term classes
 * Made the `Item` constructor parameters optional
 * Made the `Fingerprint` constructor parameters optional
+* Made the `Property` constructor parameter for fingerprint nullable
 * Deprecated `Item::newEmpty` in favour of `new Item()`
 * Deprecated `Fingerprint::newEmpty` in favour of `new Fingerprint()`
 * The `StatementList` constructor now accepts `Statement` objects in variable-lenght argument list format